@@ -10,9 +10,7 @@
     "contrib:generate": "githubcontrib --owner formly-js --repo ngx-formly --cols 6 --sortOrder desc > CONTRIBUTORS.md",
     "lint": "tslint -p tsconfig.json",
     "lint:fix": "tslint --fix -p tsconfig.json",
-    "postinstall": "patch --forward node_modules/ng-packagr/lib/ng-v5/entry-point/ts/compile-ngc.transform.js .config/ng-packagr.patch || true",
     "commit": "npm run build && git-cz",
-    "prepublish": "npm run build",
     "bundlesize": "bundlesize",
     "build": "npm run build:core && npm run build:bootstrap && npm run build:material && npm run build:ionic && npm run build:primeng && npm run build:kendo && node .config/build.js",
     "build:core": "ng-packagr -p src/core/ng-package.json",
@@ -53,23 +51,25 @@
   },
   "homepage": "https://github.com/formly-js/ngx-formly#readme",
   "dependencies": {
+    "rxjs-compat": "^6.0.0-rc.0",
     "tslib": "^1.9.0"
   },
   "devDependencies": {
-    "@angular/animations": "5.2.10",
-    "@angular/cdk": "^5.2.5",
-    "@angular/cli": "1.7.4",
-    "@angular/common": "5.2.10",
-    "@angular/compiler": "5.2.10",
-    "@angular/compiler-cli": "5.2.10",
-    "@angular/core": "5.2.10",
-    "@angular/forms": "5.2.10",
-    "@angular/http": "5.2.10",
-    "@angular/language-service": "5.2.10",
-    "@angular/material": "^5.2.5",
-    "@angular/platform-browser": "5.2.10",
-    "@angular/platform-browser-dynamic": "5.2.10",
-    "@angular/router": "5.2.10",
+    "@angular-devkit/build-angular": "~0.6.0",
+    "@angular/animations": "^6.0.0",
+    "@angular/cdk": "^6.0.0-rc.14",
+    "@angular/cli": "^6.0.0",
+    "@angular/common": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/compiler-cli": "^6.0.0",
+    "@angular/core": "^6.0.0",
+    "@angular/forms": "^6.0.0",
+    "@angular/http": "^6.0.0",
+    "@angular/language-service": "^6.0.0",
+    "@angular/material": "^6.0.0-rc.14",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.0.0",
+    "@angular/router": "^6.0.0",
     "@commitlint/cli": "^6.1.3",
     "@commitlint/config-angular": "^6.1.3",
     "@commitlint/travis-cli": "^6.1.3",
@@ -121,7 +121,7 @@
     "protractor": "~5.3.1",
     "raw-loader": "^0.5.1",
     "remap-istanbul": "^0.11.0",
-    "rxjs": "^5.5.10",
+    "rxjs": "6.1.0",
     "shx": "^0.2.2",
     "standard-version": "^4.3.0",
     "ts-node": "^6.0.0",
