@@ -26,8 +26,8 @@ import { AppComponent } from './app.component';
               { file: 'app.component.html', content: require('!!prismjs-loader?lang=html!./app.component.html'), filecontent: require('!!raw-loader?lang=html!./app.component.html') },
               { file: 'app.component.ts', content: require('!!prismjs-loader?lang=typescript!./app.component.ts'), filecontent: require('!!raw-loader?lang=typescript!./app.component.ts') },
               { file: 'user.service.ts', content: require('!!prismjs-loader?lang=typescript!./user.service.ts'), filecontent: require('!!raw-loader?lang=typescript!./user.service.ts') },
-              { file: 'user.json', content: require('!!prismjs-loader?lang=typescript!../../../../assets/json-powered/user.json'), filecontent: require('!!raw-loader?lang=typescript!../../../../assets/json-powered/user.json') },
-              { file: 'user-form.json', content: require('!!prismjs-loader?lang=typescript!../../../../assets/json-powered/user-form.json'), filecontent: require('!!raw-loader?lang=typescript!../../../../assets/json-powered/user-form.json') },
+              // { file: 'user.json', content: require('!!prismjs-loader?lang=typescript!../../../../assets/json-powered/user.json'), filecontent: require('!!raw-loader?lang=typescript!../../../../assets/json-powered/user.json') },
+              // { file: 'user-form.json', content: require('!!prismjs-loader?lang=typescript!../../../../assets/json-powered/user-form.json'), filecontent: require('!!raw-loader?lang=typescript!../../../../assets/json-powered/user-form.json') },
               { file: 'app.module.ts', content: require('!!prismjs-loader?lang=typescript!./app.module.ts'), filecontent: require('!!raw-loader?lang=typescript!./app.module.ts') },
             ],
           }],
